(fix) remove false positive warnings for unary operators
#1202
