--- packages/svelte2tsx/test/svelte2tsx/samples/uses-$store-with-exclamation-mark/input.svelte
@@ -1,7 +1,0 @@
-<script>
-  import { writable } from 'svelte/store';
-  const count = writable(0);
-  const handler1 = () => !$count
-</script>
-
-<button on:click={() => !$count}>add</button>
