+++ packages/svelte2tsx/test/svelte2tsx/samples/uses-$store-with-unary-operators/input.svelte
@@ -1,0 +1,13 @@
+<script>
+  import { writable } from 'svelte/store';
+  const count = writable(0);
+  const handler1 = () => !$count
+  const handler2 = () => +$count
+  const handler3 = () => -$count
+  const handler4 = () => ~$count
+</script>
+
+<button on:click={() => !$count}>add</button>
+<button on:click={() => +$count}>add</button>
+<button on:click={() => -$count}>add</button>
+<button on:click={() => ~$count}>add</button>
